requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.72.0rc1
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
neo4j==5.28.1
oauthlib==3.2.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
soupsieve==2.6
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
